# docs/swagger.yaml
basePath: /api
definitions:
  auth.errorResponse:
    properties:
      message:
        example: Invalid request body
        type: string
    type: object
  auth.registerRequest:
    properties:
      display_name:
        example: John Doe
        maxLength: 255
        minLength: 3
        type: string
      email:
        example: user@example.com
        maxLength: 255
        type: string
      password:
        example: password123
        maxLength: 255
        minLength: 8
        type: string
    required:
    - display_name
    - email
    - password
    type: object
  auth.registerResponse:
    properties:
      message:
        example: User registered successfully
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Ridash API server for user authentication and management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ridash API
  version: "1.0"
paths:
  /auth/oauth/{provider}:
    get:
      description: Redirects user to OAuth provider for authentication
      parameters:
      - description: OAuth provider (e.g., google, github)
        in: path
        name: provider
        required: true
        type: string
      - description: Redirect URL after successful OAuth linking
        in: query
        name: next
        type: string
      responses:
        "307":
          description: Redirect to OAuth provider
          schema:
            type: string
        "400":
          description: Invalid provider or bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Initiate OAuth flow
      tags:
      - oauth
  /auth/oauth/{provider}/callback:
    get:
      consumes:
      - application/json
      description: Handles OAuth provider callback, processes authorization code,
        creates/links user accounts, and issues authentication tokens
      parameters:
      - description: OAuth provider (e.g., google, github)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code from OAuth provider
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to success URL with authentication cookies set
          schema:
            type: string
        "400":
          description: Invalid provider, oauth state, or verification failed
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal server error during user creation or token generation
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: OAuth callback handler
      tags:
      - oauth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with email and password
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/auth.registerResponse'
        "400":
          description: Invalid request body or email already in use
          schema:
            $ref: '#/definitions/auth.errorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.errorResponse'
      summary: Register a new user
      tags:
      - auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"